Se protegieron los metodos de producto con el token

+++ NAOM-api/src/main/java/org/generation/naomdb/controller/ProductoController.java
@@ -1,12 +1,15 @@
 package org.generation.naomdb.controller;
 
 import org.generation.naomdb.exception.ProductoNotFound;
+import org.generation.naomdb.helper.TokenHelper;
 import org.generation.naomdb.model.Producto;
 import org.generation.naomdb.repository.ProductoRepository;
 import org.generation.naomdb.service.ProductoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
 import java.math.BigDecimal;
 import java.util.List;
 
@@ -37,24 +40,46 @@ public class ProductoController {
     }//getProducto
 
     @DeleteMapping(path = "{prodId}")
-    public Producto deleteProducto(@PathVariable("prodId") Long id) {
-        return productoService.deleteProducto(id);
+    public Producto deleteProducto(HttpServletRequest request,
+                                   @PathVariable("prodId") Long id) throws ServletException {
+        String token = TokenHelper.getTokenFromHeader(request);
+        if(TokenHelper.getUserRole(token).equals("admin")){
+            return productoService.deleteProducto(id);
+        }
+        throw new ServletException("Not an admin");
     }//deleteProducto
 
     @PostMapping
-    public Producto addProducto(@RequestBody Producto producto) {
-        return productoService.addProducto(producto);
+    public Producto addProducto(HttpServletRequest request,
+                                @RequestBody Producto producto) throws ServletException {
+        String token = TokenHelper.getTokenFromHeader(request);
+        if(TokenHelper.getUserRole(token).equals("admin")){
+            return productoService.addProducto(producto);
+        }
+        throw new ServletException("Not an admin");
     }//addProducto
 
     @PutMapping(path = "{prodId}")
-    public Producto updateProducto(@PathVariable("prodId") Long id,
+    public Producto updateProducto(HttpServletRequest request,
+                                   @PathVariable("prodId") Long id,
                                    @RequestParam(required = false) String nombre,
                                    @RequestParam(required = false) String descripcion,
                                    @RequestParam(required = false) String foto,
                                    @RequestParam(required = false) Double precio,
                                    @RequestParam(required = false) int stock,
                                    @RequestParam(required = false) BigDecimal rating) throws Exception {
-        return productoService.updateProducto(id, nombre, descripcion, foto, precio,
-                stock, rating);
+
+        String token = TokenHelper.getTokenFromHeader(request);
+        if(TokenHelper.getUserRole(token).equals("admin")){
+            return productoService.updateProducto(
+                    id,
+                    nombre,
+                    descripcion,
+                    foto,
+                    precio,
+                    stock,
+                    rating);
+        }
+        throw new ServletException("Not an admin");
     }//updateProducto
 }//class ProductoController
